requirements.txt
nltk==3.5
beautifulsoup4==4.9.3
scikit-learn==0.23.2
numpy==1.19.4
pyenchant==3.1.1
